requirements.txt
# S.A.M.I. - Sistema Automático de Monitoreo Inteligente
# Dependencias Python

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de Datos
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# IA y Machine Learning
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1
opencv-python==4.8.1.78
face-recognition==1.3.0
whisper==1.1.10
pyttsx3==2.90

# Procesamiento de Audio/Video
pyaudio==0.2.11
sounddevice==0.4.6
librosa==0.10.1
moviepy==1.0.3

# Comunicación y APIs
twilio==8.10.3
requests==2.31.0
websockets==12.0

# Reportes y Documentos
reportlab==4.0.7
openpyxl==3.1.2
jinja2==3.1.2

# Utilidades
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
pytz==2023.3

# Hardware y IoT
pyserial==3.5
RPi.GPIO==0.7.1
adafruit-circuitpython-rfid==1.0.0

# Monitoreo y Logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Desarrollo
black==23.11.0
flake8==6.1.0
mypy==1.7.1
